Changed IntegerType to use 64 bit int by default.

+++ core/modules/core/types/integer_type.php
@@ -2,7 +2,7 @@
 
 class IntegerType extends \nmvc\AppType {
     public function getSQLType() {
-        return "int";
+        return "bigint";
     }
 
     public function get() {
